[M2-002] Code Review Result
