+++ app/src/main/java/id/dev/qrcraft/navigation/navs/AppNavigation.kt
@@ -4,12 +4,10 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.consumeWindowInsets
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavHostController
-import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import androidx.navigation.navArgument
 import id.dev.core.presentation.R
 import id.dev.home.presentation.camera.CameraScreenRoot
 import id.dev.home.presentation.create_qr.CreateQRRoot
@@ -25,12 +23,14 @@ fun AppNavigation(
     navController: NavHostController,
     contentPadding: PaddingValues
 ) {
+    val context = LocalContext.current
+
     NavHost(
         navController = navController,
-        startDestination = Screens.CameraScreen.route,
+        startDestination = Screens.CameraScreen,
         modifier = Modifier.consumeWindowInsets(contentPadding)
     ) {
-        composable(Screens.CameraScreen.route) {
+        composable<Screens.CameraScreen> {
             CameraScreenRoot(
                 onScanResult = { barcodeResult ->
                     if (barcodeResult !is QrTypes.Error) {
@@ -38,44 +38,34 @@ fun AppNavigation(
                         navController.navigate(
                             route = Screens.ScanResultScreen(
                                 qrTypes = result,
-                                titleVal = "Scan Result"
-                            ).route
+                                titleVal = context.getString(R.string.scan_result)
+                            )
                         )
                     }
                 },
             )
         }
-        composable(
-            route = Screens.ScanResultScreen.ROUTE_PATTERN,
-            arguments = listOf(
-                navArgument("qrTypes") { type = NavType.StringType },
-                navArgument("titleVal") { type = NavType.StringType }
-            )
-        ) { backStackEntry ->
+        composable<Screens.ScanResultScreen> { backStackEntry ->
             ScanResultScreenRoot(
                 onNavigateUp = {
                     navController.navigateUp()
                 },
             )
         }
-        composable(Screens.CreateQrScreen.route) {
+        composable<Screens.CreateQrScreen> {
             CreateQRRoot(
                 onNavigateToGenerator = { qrTypeIdentifier ->
-                    val result = Json.encodeToString(QrTypeIdentifier.serializer(), qrTypeIdentifier)
+                    val result =
+                        Json.encodeToString(QrTypeIdentifier.serializer(), qrTypeIdentifier)
                     navController.navigate(
                         route = Screens.GenerateQrScreen(
                             qrType = result
-                        ).route
+                        )
                     )
                 }
             )
         }
-        composable(
-            route = Screens.GenerateQrScreen.ROUTE_PATTERN,
-            arguments = listOf(
-                navArgument("qrType") { type = NavType.StringType }
-            )
-        ) { backStackEntry ->
+        composable<Screens.GenerateQrScreen> { backStackEntry ->
             CreateQrGeneratorRoot(
                 onNavigateUp = {
                     navController.navigateUp()
@@ -84,8 +74,8 @@ fun AppNavigation(
                     navController.navigate(
                         route = Screens.ScanResultScreen(
                             qrTypes = qr,
-                            titleVal = "Preview"
-                        ).route
+                            titleVal = context.getString(R.string.preview)
+                        )
                     )
                 }
             )
