+++ app/src/main/java/id/dev/qrcraft/navigation/screens/Screens.kt
@@ -3,39 +3,25 @@ package id.dev.qrcraft.navigation.screens
 import kotlinx.serialization.Serializable
 
 @Serializable
-sealed class Screens(open val route: String) {
+sealed interface Screens {
 
     @Serializable
-    data object CameraScreen : Screens(route = "camera_screen")
+    data object CameraScreen : Screens
 
     @Serializable
     data class ScanResultScreen(
         val qrTypes: String,
         val titleVal: String
-    ) : Screens("scan_result_screen") {
-        override val route: String
-            get() = "scan_result_screen/$qrTypes/$titleVal"
-
-        companion object {
-            const val ROUTE_PATTERN = "scan_result_screen/{qrTypes}/{titleVal}"
-        }
-    }
+    ) : Screens
 
     @Serializable
-    data object CreateQrScreen : Screens("create_qr_screen")
+    data object CreateQrScreen : Screens
 
     @Serializable
     data class GenerateQrScreen(
         val qrType: String
-    ) : Screens("generate_qr_screen") {
-        override val route: String
-            get() = "generate_qr_screen/$qrType"
-
-        companion object {
-            const val ROUTE_PATTERN = "generate_qr_screen/{qrType}"
-        }
-    }
+    ) : Screens
 
     @Serializable
-    data object HistoryQrScreen : Screens("history_qr_screen")
+    data object HistoryQrScreen : Screens
 }
